requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
azure-core==1.37.0
azure-identity==1.25.1
banks==2.2.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.0
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
duckdb==1.4.3
et_xmlfile==2.0.0
eval_type_backport==0.3.1
fastavro==1.12.1
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
griffe==1.15.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
idna==3.11
invoke==2.2.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.10
llama-index-cli==0.5.3
llama-index-core==0.14.10
llama-index-embeddings-adapter==0.4.1
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-experimental==0.6.3
llama-index-finetuning==0.4.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-azure-openai==0.4.2
llama-index-llms-groq==0.4.1
llama-index-llms-openai==0.6.10
llama-index-llms-openai-like==0.5.3
llama-index-postprocessor-cohere-rerank==0.5.1
llama-index-readers-file==0.5.5
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.11.5
llama-parse==0.6.54
MarkupSafe==3.0.3
marshmallow==3.26.1
mistralai==1.9.11
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.13.0
nest-asyncio==1.6.0
networkx==3.6
nltk==3.9.2
numpy==2.3.5
openai==2.9.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==10.4.0
platformdirs==4.5.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.10.1
pyngrok==7.5.0
pyparsing==3.2.5
pypdf==6.4.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
streamlit==1.52.1
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.3
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.1
watchdog==6.0.0
wrapt==1.17.3
yarl==1.22.0
